requirements.txt
Django==2.2
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
pkg-resources==0.0.0
psycopg2==2.8.2
PyJWT==1.7.1
pytz==2019.1
sqlparse==0.3.0
